refactor(wiring): move wire proposals into edit engine
