+++ packages/edit-engine/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./history.js";
+export * from "./routing-planner.js";
 export * from "./transaction.js";
